refactor(multisig): address Davi0k re-review — renames, map returns, error-tag revert, owner-mgmt
IWalletAccountReadOnlyBase→IWalletAccountReadOnlySimple (class+file); getMessages→getMessageProposals; approveMessage→approveMessageProposal; getProposals/getMessageProposals now return Record<string, X|null> maps; signerKeyPair→keyPair; @throws {SignerError}→{Error} (PR #54 coordination); rename owner-management file (drop i- prefix) + simplify its interface doc. Deferred pending reviewer: message-type dedup (Comments 2/3), single-item getters (8), and open questions (11/12/13).
