+++ src/multisig/wallet-account-multisig.js
@@ -24,7 +24,6 @@ import { NotImplementedError } from '../errors.js'
 
 /** @typedef {import('./wallet-account-read-only-multisig.js').MultisigProposal} MultisigProposal */
 
-/** @typedef {import('../errors.js').SignerError} SignerError */
 /** @typedef {import('../errors.js').NoSuchElementError} NoSuchElementError */
 /** @typedef {import('../errors.js').ValueError} ValueError */
 
@@ -67,8 +66,8 @@ export class IWalletAccountMultisig extends IWalletAccountReadOnlyMultisig {
    *
    * @type {KeyPair}
    */
-  get signerKeyPair () {
-    throw new NotImplementedError('signerKeyPair')
+  get keyPair () {
+    throw new NotImplementedError('keyPair')
   }
 
   /**
@@ -79,7 +78,7 @@ export class IWalletAccountMultisig extends IWalletAccountReadOnlyMultisig {
    * @param {Transaction} tx - The transaction.
    * @param {MultisigTransactionOptions} [transactionOptions] - The multisig transaction's options.
    * @returns {Promise<MultisigProposal>} The created proposal; its `status` is `'executed'` when `autoExecute` ran to completion, otherwise `'pending'`.
-   * @throws {SignerError} If the signer is not an owner of the multisig account.
+   * @throws {Error} If the signer is not an owner of the multisig account.
    */
   async propose (tx, transactionOptions) {
     throw new NotImplementedError('propose(tx, transactionOptions)')
@@ -90,7 +89,7 @@ export class IWalletAccountMultisig extends IWalletAccountReadOnlyMultisig {
    *
    * @param {string} message - The message to sign.
    * @returns {Promise<MultisigMessageProposal>} The multisig message proposal.
-   * @throws {SignerError} If the signer is not an owner of the multisig account.
+   * @throws {Error} If the signer is not an owner of the multisig account.
    */
   async proposeMessage (message) {
     throw new NotImplementedError('proposeMessage(message)')
@@ -101,11 +100,11 @@ export class IWalletAccountMultisig extends IWalletAccountReadOnlyMultisig {
    *
    * @param {string} messageId - The message's hash.
    * @returns {Promise<MultisigMessageProposal>} The multisig message proposal.
-   * @throws {SignerError} If the signer is not an owner of the multisig account.
+   * @throws {Error} If the signer is not an owner of the multisig account.
    * @throws {NoSuchElementError} If no message exists for the given id.
    */
-  async approveMessage (messageId) {
-    throw new NotImplementedError('approveMessage(messageId)')
+  async approveMessageProposal (messageId) {
+    throw new NotImplementedError('approveMessageProposal(messageId)')
   }
 
   /**
@@ -113,7 +112,7 @@ export class IWalletAccountMultisig extends IWalletAccountReadOnlyMultisig {
    *
    * @param {string} proposalId - The proposal's id.
    * @returns {Promise<MultisigProposal>} The multisig proposal.
-   * @throws {SignerError} If the signer is not an owner of the multisig account.
+   * @throws {Error} If the signer is not an owner of the multisig account.
    * @throws {NoSuchElementError} If no proposal exists for the given id.
    */
   async approveProposal (proposalId) {
