+++ types/src/multisig/multisig-owner-management.d.ts
@@ -1,8 +1,5 @@
 /**
- * Optional owner-management surface for multisig accounts whose owner set is mutable
- * (e.g. account-abstraction wallets). Chains whose owner set is fixed at creation —
- * such as Bitcoin script multisig, where the participants are committed in the redeem
- * script — do not implement this interface.
+ * Adds owner management features to a multisig wallet.
  *
  * @interface
  */
@@ -13,7 +10,7 @@ export interface IMultisigOwnerManagement {
      * @param {string} owner - The owner's address.
      * @param {MultisigOptions} [options] - The multisig options.
      * @returns {Promise<MultisigProposal>} The multisig proposal.
-     * @throws {SignerError} If the signer is not an owner of the multisig account.
+     * @throws {Error} If the signer is not an owner of the multisig account.
      */
     addOwner(owner: string, options?: MultisigOptions): Promise<MultisigProposal>;
     /**
@@ -22,7 +19,7 @@ export interface IMultisigOwnerManagement {
      * @param {string} owner - The owner's address.
      * @param {MultisigOptions} [options] - The multisig options.
      * @returns {Promise<MultisigProposal>} The multisig proposal.
-     * @throws {SignerError} If the signer is not an owner of the multisig account.
+     * @throws {Error} If the signer is not an owner of the multisig account.
      */
     removeOwner(owner: string, options?: MultisigOptions): Promise<MultisigProposal>;
     /**
@@ -31,7 +28,7 @@ export interface IMultisigOwnerManagement {
      * @param {string} oldOwner - The old owner.
      * @param {string} newOwner - The new owner.
      * @returns {Promise<MultisigProposal>} The multisig proposal.
-     * @throws {SignerError} If the signer is not an owner of the multisig account.
+     * @throws {Error} If the signer is not an owner of the multisig account.
      */
     swapOwner(oldOwner: string, newOwner: string): Promise<MultisigProposal>;
     /**
@@ -39,7 +36,7 @@ export interface IMultisigOwnerManagement {
      *
      * @param {number} newThreshold - The new threshold.
      * @returns {Promise<MultisigProposal>} The multisig proposal.
-     * @throws {SignerError} If the signer is not an owner of the multisig account.
+     * @throws {Error} If the signer is not an owner of the multisig account.
      */
     changeThreshold(newThreshold: number): Promise<MultisigProposal>;
 }
